Tests de los enemigos

+++ src/server_src/personaje/personaje.h
@@ -36,7 +36,7 @@ protected:
     uint32_t id_jugador;
     uint32_t posicion_x;
     uint32_t posicion_y;
-    int vida; //DEBE SER INT PARA PODER SER NEGATIVO
+    int8_t vida; //DEBE SER INT PARA PODER SER NEGATIVO
     bool esta_quieto;
 
     int velocidad_x;
